+++ ccode/example.html
@@ -5,5 +5,5 @@
     <!-- <link rel="stylesheet" type="text/css" href="styles.css"> -->
   </head>
   <body>
-<h1>hola1</h1>
+<h1>output en consola</h1>
 </html>
